migrate group metadata logic to its own module
