+++ src/zarr/core/array.py
@@ -66,6 +66,7 @@ from zarr.core.common import (
     product,
 )
 from zarr.core.config import config as zarr_config
+from zarr.core.group import AsyncGroup
 from zarr.core.indexing import (
     BasicIndexer,
     BasicSelection,
@@ -98,6 +99,7 @@ from zarr.core.metadata import (
     ArrayV2MetadataDict,
     ArrayV3Metadata,
     ArrayV3MetadataDict,
+    GroupMetadata,
     T_ArrayMetadata,
 )
 from zarr.core.metadata.v2 import (
@@ -123,7 +125,6 @@ if TYPE_CHECKING:
 
     from zarr.abc.codec import CodecPipeline
     from zarr.codecs.sharding import ShardingCodecIndexLocation
-    from zarr.core.group import AsyncGroup
     from zarr.storage import StoreLike
 
 
@@ -3738,8 +3739,6 @@ async def chunks_initialized(
 def _build_parents(
     node: AsyncArray[ArrayV2Metadata] | AsyncArray[ArrayV3Metadata] | AsyncGroup,
 ) -> list[AsyncGroup]:
-    from zarr.core.group import AsyncGroup, GroupMetadata
-
     store = node.store_path.store
     path = node.store_path.path
     if not path:
