+++ tests/test_metadata/test_v2.py
@@ -10,8 +10,7 @@ import zarr.api.asynchronous
 import zarr.storage
 from zarr.core.buffer import cpu
 from zarr.core.buffer.core import default_buffer_prototype
-from zarr.core.group import ConsolidatedMetadata, GroupMetadata
-from zarr.core.metadata import ArrayV2Metadata
+from zarr.core.metadata import ArrayV2Metadata, ConsolidatedMetadata, GroupMetadata
 from zarr.core.metadata.v2 import parse_zarr_format
 
 if TYPE_CHECKING:
